tests/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django_filter==24.3
drf_spectacular==0.27.2
pytest-django==4.9.0
pytest-factoryboy==2.7.0
pytest-cov==5.0.0